+++ CHANGES.rst
@@ -28,6 +28,9 @@ Release 2.4.1 (2015-11-25)
 
 - Add support for viewing, adding and deleting annotations on assets.
 
+- Remove htmlmin from build it messes up the html and is not necessary with
+  gzip.
+
 
 Release 2.3.2 (2015-11-11)
 ---------------------