+++ CHANGES.rst
@@ -3,7 +3,13 @@ Changelog of lizard-nxt client
 
 Unreleased (2.2.7) (XXXX-XX-XX)
 ---------------------
--
+
+- Fix click on animation pause button not registered.
+
+- Animate only the intersection of map bounds and layer bounds, to have more
+  resolution with less data.
+
+- Store bounds of layer on group and layer for zooming to lg and animating wms.
 
 
 Release 2.2.7 (2015-11-25)