+++ CHANGES.rst
@@ -3,6 +3,11 @@ Changelog of lizard-nxt client
 
 Unreleased (2.2.7) (XXXX-XX-XX)
 ---------------------
+-
+
+
+Release 2.2.8 (2015-12-7)
+---------------------
 
 - Fix click on animation pause button not registered.